+++ crates/dxm-manifest/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `artifact::SYSTEM_RESOURCES` constant.
 - `artifact::TXADMIN_MONITOR` constant.
 - Optional third-party monitor in artifact and lockfile definitions.
+- Profile structures.
 
 ### Changed